# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Publisher Author Marketplace API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/matches/run": {
      "post": {
        "operationId": "matches_run",
        "tags": ["matches"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "run": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "direction": {
                          "type": "string",
                          "enum": [
                            "author_to_publisher",
                            "publisher_to_manuscript"
                          ]
                        },
                        "requesterProfileId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "sourceManuscriptId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "sourcePublisherProfileId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "status": {
                          "type": "string",
                          "enum": ["running", "succeeded", "failed"]
                        },
                        "stale": {
                          "type": "boolean"
                        },
                        "candidateCount": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "failureCode": {
                          "type": "string",
                          "maxLength": 80,
                          "nullable": true
                        },
                        "inputFingerprint": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "sourceTitle": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "direction",
                        "requesterProfileId",
                        "sourceManuscriptId",
                        "sourcePublisherProfileId",
                        "status",
                        "stale",
                        "candidateCount",
                        "failureCode",
                        "inputFingerprint",
                        "sourceTitle",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "candidates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "runId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "rank": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "candidateProfileId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "candidateManuscriptId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "candidateType": {
                            "type": "string",
                            "enum": ["publisher", "manuscript"]
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "subtitle": {
                            "type": "string",
                            "maxLength": 200,
                            "nullable": true
                          },
                          "scoreBand": {
                            "type": "string",
                            "enum": ["strong", "moderate", "weak"]
                          },
                          "axisBands": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string",
                              "enum": ["strong", "moderate", "weak"]
                            }
                          },
                          "explanation": {
                            "type": "string",
                            "maxLength": 1200,
                            "nullable": true
                          },
                          "fitReasons": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 240
                            },
                            "maxItems": 8
                          },
                          "riskReasons": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 240
                            },
                            "maxItems": 8
                          },
                          "profilePath": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 240
                          },
                          "manuscriptProfilePath": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 240,
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "runId",
                          "rank",
                          "candidateProfileId",
                          "candidateManuscriptId",
                          "candidateType",
                          "title",
                          "subtitle",
                          "scoreBand",
                          "axisBands",
                          "explanation",
                          "fitReasons",
                          "riskReasons",
                          "profilePath",
                          "manuscriptProfilePath"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["run", "candidates"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches": {
      "get": {
        "operationId": "matches_list",
        "tags": ["matches"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "direction": {
                            "type": "string",
                            "enum": [
                              "author_to_publisher",
                              "publisher_to_manuscript"
                            ]
                          },
                          "requesterProfileId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "sourceManuscriptId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "sourcePublisherProfileId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": ["running", "succeeded", "failed"]
                          },
                          "stale": {
                            "type": "boolean"
                          },
                          "candidateCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "failureCode": {
                            "type": "string",
                            "maxLength": 80,
                            "nullable": true
                          },
                          "inputFingerprint": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 128
                          },
                          "sourceTitle": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "direction",
                          "requesterProfileId",
                          "sourceManuscriptId",
                          "sourcePublisherProfileId",
                          "status",
                          "stale",
                          "candidateCount",
                          "failureCode",
                          "inputFingerprint",
                          "sourceTitle",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["runs"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/history": {
      "get": {
        "operationId": "matches_profileHistory",
        "tags": ["matches"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "direction": {
                            "type": "string",
                            "enum": [
                              "author_to_publisher",
                              "publisher_to_manuscript"
                            ]
                          },
                          "requesterProfileId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "sourceManuscriptId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "sourcePublisherProfileId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": ["running", "succeeded", "failed"]
                          },
                          "stale": {
                            "type": "boolean"
                          },
                          "candidateCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "failureCode": {
                            "type": "string",
                            "maxLength": 80,
                            "nullable": true
                          },
                          "inputFingerprint": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 128
                          },
                          "sourceTitle": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "direction",
                          "requesterProfileId",
                          "sourceManuscriptId",
                          "sourcePublisherProfileId",
                          "status",
                          "stale",
                          "candidateCount",
                          "failureCode",
                          "inputFingerprint",
                          "sourceTitle",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["runs"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{matchRunId}": {
      "get": {
        "operationId": "matches_get",
        "tags": ["matches"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "matchRunId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "run": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "direction": {
                          "type": "string",
                          "enum": [
                            "author_to_publisher",
                            "publisher_to_manuscript"
                          ]
                        },
                        "requesterProfileId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "sourceManuscriptId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "sourcePublisherProfileId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "status": {
                          "type": "string",
                          "enum": ["running", "succeeded", "failed"]
                        },
                        "stale": {
                          "type": "boolean"
                        },
                        "candidateCount": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "failureCode": {
                          "type": "string",
                          "maxLength": 80,
                          "nullable": true
                        },
                        "inputFingerprint": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "sourceTitle": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "direction",
                        "requesterProfileId",
                        "sourceManuscriptId",
                        "sourcePublisherProfileId",
                        "status",
                        "stale",
                        "candidateCount",
                        "failureCode",
                        "inputFingerprint",
                        "sourceTitle",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "candidates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "runId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "rank": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "candidateProfileId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "candidateManuscriptId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "candidateType": {
                            "type": "string",
                            "enum": ["publisher", "manuscript"]
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "subtitle": {
                            "type": "string",
                            "maxLength": 200,
                            "nullable": true
                          },
                          "scoreBand": {
                            "type": "string",
                            "enum": ["strong", "moderate", "weak"]
                          },
                          "axisBands": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string",
                              "enum": ["strong", "moderate", "weak"]
                            }
                          },
                          "explanation": {
                            "type": "string",
                            "maxLength": 1200,
                            "nullable": true
                          },
                          "fitReasons": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 240
                            },
                            "maxItems": 8
                          },
                          "riskReasons": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 240
                            },
                            "maxItems": 8
                          },
                          "profilePath": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 240
                          },
                          "manuscriptProfilePath": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 240,
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "runId",
                          "rank",
                          "candidateProfileId",
                          "candidateManuscriptId",
                          "candidateType",
                          "title",
                          "subtitle",
                          "scoreBand",
                          "axisBands",
                          "explanation",
                          "fitReasons",
                          "riskReasons",
                          "profilePath",
                          "manuscriptProfilePath"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["run", "candidates"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-webhook-signature"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {}
            },
            "required": ["code", "message"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      }
    }
  }
}